Stopwatch added with controller and view are on seperated pages
