+++ lib/app/modules/stopwatch/views/stopwatch_view.dart
@@ -1,55 +1,16 @@
-import 'dart:async';
-
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:ultimate_alarm_clock/app/modules/settings/controllers/theme_controller.dart';
-import 'package:ultimate_alarm_clock/app/modules/timer/controllers/timer_controller.dart';
-
+import 'package:ultimate_alarm_clock/app/modules/stopwatch/controllers/stopwatch_controller.dart';
 import '../../../utils/constants.dart';
 import '../../../utils/utils.dart';
+import '../controllers/stopwatch_controller.dart';
 
-class StopwatchView extends GetView<TimerController> {
+class StopwatchView extends GetView<StopwatchController> {
   StopwatchView({Key? key}) : super(key: key);
 
   ThemeController themeController = Get.find<ThemeController>();
-  final Stopwatch _stopwatch = Stopwatch();
-  late Timer timer;
-  RxBool isTimerPaused = true.obs;
-
-  RxString _result = '00:00:00'.obs;
-
-  void _toggleTimer() {
-    if (isTimerPaused.value) {
-      _startTimer();
-    } else {
-      _stopTimer();
-    }
-  }
-
-  void _startTimer() {
-    timer = Timer.periodic(const Duration(milliseconds: 30), (Timer t) {
-      _updateResult();});
-    _stopwatch.start();
-    isTimerPaused.value = false;
-  }
-
-  void _stopTimer() {
-    timer.cancel();
-    _stopwatch.stop();
-    isTimerPaused.value = true;
-  }
-
-  void _resetTime() {
-    _stopTimer();
-    _stopwatch.reset();
-    _updateResult();
-  }
-
-  void _updateResult() {
-    _result.value =
-    '${_stopwatch.elapsed.inMinutes.toString().padLeft(2, '0')}:${(_stopwatch.elapsed.inSeconds % 60).toString().padLeft(2, '0')}:${(_stopwatch.elapsed.inMilliseconds % 100).toString().padLeft(2, '0')}';
-  }
-
+  StopwatchController stopwatchController=Get.put(StopwatchController());
   @override
   Widget build(BuildContext context) {
     var width = Get.width;
@@ -65,7 +26,6 @@ class StopwatchView extends GetView<TimerController> {
             IconButton(
               onPressed: () {
                 Utils.hapticFeedback();
-                controller.saveTimerStateToStorage();
                 Get.toNamed('/settings');
               },
               icon: const Icon(
@@ -84,11 +44,9 @@ class StopwatchView extends GetView<TimerController> {
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
             Obx(() => Text(
-              _result.value,
+              controller.result,
               style: const TextStyle(
-                fontSize: 60.0,
-                fontWeight: FontWeight.bold
-              ),
+                  fontSize: 60.0, fontWeight: FontWeight.bold),
             )),
             const SizedBox(
               height: 20.0,
@@ -97,17 +55,18 @@ class StopwatchView extends GetView<TimerController> {
               mainAxisAlignment: MainAxisAlignment.spaceEvenly,
               children: [
                 FloatingActionButton(
-                  onPressed: _toggleTimer,
+                  onPressed: controller.toggleTimer,
                   child: Obx(() => Icon(
-                    isTimerPaused.value ? Icons.play_arrow : Icons.pause,
+                    controller.isTimerPaused.value
+                        ? Icons.play_arrow
+                        : Icons.pause,
                   )),
                 ),
                 // Reset button
                 FloatingActionButton(
-                  onPressed: _resetTime,
-                  child: Icon(Icons.square_rounded)
-                  )
-
+                  onPressed: controller.resetTime,
+                  child: Icon(Icons.square_rounded),
+                ),
               ],
             ),
           ],
